requirements.txt
crewai==0.28.0
duckduckgo-search==4.1.0
python-dotenv==1.0.0
